requirements.txt
Flask >= 1.1.1
ShopifyAPI >= 5.1.2
python-dotenv >= 0.10.3